Page object model implemented with FindBy annotations.
