+++ src/main/java/pages/HomePage.java
@@ -1,15 +1,20 @@
 package pages;
 
-import tools.ObjectActions;
+import io.appium.java_client.MobileElement;
+import org.openqa.selenium.support.FindBy;
 import tools.ObjectVisibility;
 import tools.Scroller;
 
 
 public class HomePage {
 
+    //Page object model example
+    @FindBy(id = "btnAccept")
+    private static MobileElement readAndAcceptBtn;
+
     public static void userOnHomePage(){
         try {
-            ObjectActions.set("UiReadAndAccept", "LCLICK", null, null);
+            readAndAcceptBtn.click();
             ObjectVisibility.assertInVisibility("UiReadAndAccept","DEFAULT", null);
         } catch (Exception e){
             System.out.println("Error!! Application home page is not visible!");
