+++ src/test/java/stepDefinitions/AddProductToCart.java
@@ -1,31 +1,14 @@
 package stepDefinitions;
 
-import register.Driver;
-import tools.AppSpecificActions;
-import io.cucumber.java.AfterStep;
-import io.cucumber.java.BeforeAll;
-import io.cucumber.java.Scenario;
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
-import org.openqa.selenium.OutputType;
-import org.openqa.selenium.TakesScreenshot;
+
 import stepMethodDecorators.AddProductToCartDecorator;
 
 
 public class AddProductToCart {
-
-    @BeforeAll
-    public static void before_all() {
-        AppSpecificActions.lunchApp();
-    }
-
-    @AfterStep
-    public void addScreenshot(Scenario scenario){
-        final byte[] screenshot = ((TakesScreenshot) Driver.getDriver()).getScreenshotAs(OutputType.BYTES);
-        scenario.attach(screenshot, "image/png", "image");
-    }
 
     @Given("^Mike on home page after opening nopCart mobile app$")
     public void mike_on_home_page_after_opening_nop_cart_mobile_app() {
